+++ src/components/SettingsControl/components/index.js
@@ -1,0 +1,3 @@
+export { default as SettingsPopover } from './SettingsPopover';
+export { default as SettingsContent } from './SettingsContent';
+export { default as FlagIcon } from './FlagIcon';