+++ src/locales/en/common.js
@@ -1,4 +1,6 @@
 const common = {
+  language: 'Language',
+  languageName: 'English',
   yes: 'Yes',
   no: 'No',
   cancel: 'Cancel',