+++ src/locales/pl/common.js
@@ -1,4 +1,6 @@
 const common = {
+  language: 'Język',
+  languageName: 'polski',
   yes: 'Tak',
   no: 'Nie',
   cancel: 'Anuluj',